fix race condition for waitComplete
this makes the code look a little silly, but this should actually be
the same behavior except that it prevents a race condition where two
threads could have just triggered sent in send() and then passed the
sent guard in waitComplete() but prior to either thread acquiring a
permit from the semaphore. It does indicate a need to rethink the
rest of what's going on - for instance, do we really need to check
the sent flag in receiveComplete() before releasing the gate? I'm
not waitComplete()'s main goal is or what it's warning comment refers
to so i'll leave it relatively alone for now. Meanwhile, the sent
flag may be just what I need for something else.