fix field initialization in fowarded file targets
from the added comment (about doComplete):

"""
must be initialized in init() because (unfortunately) it may be required before
the parent class's constructor finishes. init() is called from the parent class's
constructor; allowing us to hack in some sub-class construction. This is why
constructors should do as little as possible, and really, _really_ not publish
incomplete objects to other threads.
"""