try to handle spurious frames in stream target
this mostly re-solves the extra 'send more' frames that can be
sent after a target self-completes and removes itself from the
handler map. The creation frames logic was intended to handle
this, but since it causes problems with older versions and this
type of issue may possibly occur in another way, it is fine to
be a little more robust.