handle bad remote mesh uuids/searches et al
a while ago, StreamSource was changed to throw an exception during
construction (least of two evils) when no requested mesh nodes are
found. StreamTarget's proxy stream source was not handling that
possible exception and was instead still attempting to check for
that condition manually. The exception was handled in a less useful
way by one of the fail safe handlers, so this restores the desired
behavior.

in particular, SourceInputStream does not seem to handle unexpected,
early session completion in a clean manner. While that is somewhat
its own problem, this issue was causing endless waits for blocking
reads.