add per-channel pending write limit fail-safes
this should mostly come into play for things like file finds which
are otherwise done as blind writes and can topple mesh nodes if
hundreds of thousands of files are able to be found extremely quickly.

I believe there are some nasty edge cases that can form from ending
up in a "write-only" situation where you are stuck there until reads
are done elsewhere -- and they turn out to be in the same boat and
so on. However:

- those cases are likely to have OOMed before
- the high default watermark should cause stream sources to be more
  or less left alone (unless they use a high custom buffer)
- if the origin of the slowdown is a terminus / client, then the
  blocking should always eventually clear as it catches up and frees
  io threads recursively. In practice, this seems to be the most
  common cause.
- I put in an escape hatch just in case

a better solution might be to implement windowing for file sources
similar to stream sources, but that involves a lot of changes to
the protocol and seems more error prone.