fixes for both under and over peering + logging
the "allChannels" field has been removed for being redundant with
"connectedChannels" and the "mC" metric that counted its size has
been switched to the size of "connectedChannels". That used to be
"sC", which is now the actual number of peers rather than peers +
true clients + over peering.

under peering could occur with multiple different servers connected
to the same additional server when they relied on that shared server
to introduce them to each other. This was a race condition caused by
sending a peering request and a list of current peers at the same
time (so two requests could go out and neither's list would include
the other). Now peering happens as a discrete step prior to getting
to sending a list of peers. This should resolve some random test
failures.

over peering has been a problem for a while, and although somewhat
harmless, it does tend to waste resources and throw off both tests
and metrics. It has been more or less (entirely?) resolved by a
strategy that randomly but deterministically, and without any extra
negotiation, decides which of two servers should be the one who
creates the client connection to the other's listening port. This
resolves the case where two servers are racing to peer with each
other. There is an additional case where a server ends up creating
multiple connections / peer requests by itself, and this is fixed
by the non-client / listening server promoting peers in an atomic
fashion and rejecting additional peer requests (this also relies
on the client-server decision strategy and cannot be done on the
connecting server because it does not have the remote uuid yet).

additionally, there is a fix for failed client connections never
being allowed to retry (getting stuck in "inPeering"), remote
addresses not being correctly narrowed to an actual IP, and maybe
some other stuff.