give up final fields to avoid channelUnregistered
channelUnregistered is where the bug is in 26, and apparently it is
being removed in future versions. That probably partially explains
why it was not super well tested. Given these facts, let's just go
ahead and distance ourselves from it. I guess this means we can also
use 26 now.. probably.. so let's go ahead and do that.